[llvm-cov] Add one correction to r315960 (PR34962)
In r315960, I accidentally assumed that the first line segment is
guaranteed to be the non-gap region entry segment (given that one is
present). It can actually be any segment on the line, and the test I
checked in demonstrates that.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@315963 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 5128bb371a5f35197e5cd7f66909d2447358e4e7)

apple-llvm-split-commit: cfb29a2de68df774308b4bac83ab5b19de71b249
apple-llvm-split-dir: llvm/