[llvm-cov] Move LineCoverageIterator to libCoverage. NFC.
LineCoverageIterator makes it easy for clients of coverage data to
determine line execution counts for a file or function. The coverage
iteration logic is tricky enough that it really pays not to have
multiple copies of it. Hopefully having just one implementation in LLVM
will make the iteration logic easier to test, reuse, and update.

This commit is NFC but I've added a unit test to go along with it just
because it's easy to do now.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@316141 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 2ce33a5dffc95368fb7d6faa5e6f4b9b5471c3b9)

apple-llvm-split-commit: 14c235de73a142e902157d4f9d8c20a6a7f5ae34
apple-llvm-split-dir: llvm/