MIR: Print the register class or bank in vreg defs
This updates the MIRPrinter to include the regclass when printing
virtual register defs, which is already valid syntax for the
parser. That is, given 64 bit %0 and %1 in a "gpr" regbank,

  %1(s64) = COPY %0(s64)

would now be written as

  %1:gpr(s64) = COPY %0(s64)

While this change alone introduces a bit of redundancy with the
registers block, it allows us to update the tests to be more concise
and understandable and brings us closer to being able to remove the
registers block completely.

Note: We generally only print the class in defs, but there is one
exception. If there are uses without any defs whatsoever, we'll print
the class on all uses. I'm not completely convinced this comes up in
meaningful machine IR, but for now the MIRParser and MachineVerifier
both accept that kind of stuff, so we don't want to have a situation
where we can print something we can't parse.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@316479 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit edab7579664b13faad3680c676c6b8a49d2ea00b)

Conflicts:
      test/CodeGen/AArch64/GlobalISel/call-translator.ll
      test/CodeGen/AArch64/GlobalISel/debug-insts.ll
      test/CodeGen/AMDGPU/endpgm-dce.mir
      test/CodeGen/AMDGPU/macro-fusion-cluster-vcc-uses.mir
      test/CodeGen/AMDGPU/shrink-vop3-carry-out.mir
      test/CodeGen/AMDGPU/twoaddr-mad.mir
      test/CodeGen/ARM/GlobalISel/arm-select-globals-ropi-rwpi.mir
      test/CodeGen/Hexagon/cext-opt-basic.mir
      test/CodeGen/Hexagon/early-if-debug.mir
      test/CodeGen/Hexagon/hwloop-redef-imm.mir
      test/CodeGen/MIR/X86/unreachable-mbb-undef-phi.mir
      test/CodeGen/X86/GlobalISel/irtranslator-callingconv.ll
      test/CodeGen/X86/debugloc-no-line-0.ll
      test/CodeGen/X86/domain-reassignment.mir
      test/CodeGen/X86/movtopush.mir

apple-llvm-split-commit: 09bd7480cae295a7dd94c1490398173c0909c61c
apple-llvm-split-dir: llvm/