[ARM GlobalISel] Support G_FDIV for s32 and s64
TableGen already generates code for selecting a G_FDIV, so we only need
to add a test.

For the legalizer and reg bank select, we do the same thing as for the
other floating point binary operations: either mark as legal if we have
a FP unit or lower to a libcall, and map to the floating point
registers.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@318915 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit ffb04cb1776fad962bf542ec0c341b3bb38a44c4)

apple-llvm-split-commit: 8e66248bbe89032e23695df95f9fa5ad83ef79a1
apple-llvm-split-dir: llvm/