Change the ordering that we search for kexts and kernels on the local computer.  When doing kernel debugging, lldb scrapes around a few well-known locations to find kexts and kernels.  It builds up two lists - kexts and kernels with dSYM, and kexts and kernels without dSYMs. After both lists have failed to provide a file, then we'll call out to things like the DebugSymbols framework to find a kext/kernel.
This meant that when you had a kext/kernel on the local computer that
did not have debug information, lldb wouldn't consult DebugSymbols etc
once it'd locked on to one of these no-debug-info binaries on the local
computer.

Reorder this so we give DebugSymbols etc a shot at finding a debug-info
file before we use any of the no-debug-info binaries that were found on
the system.

<rdar://problem/34434440>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@320241 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 95b94f47d04caf93a93fff7f4ac76092ec362a68)

apple-llvm-split-commit: e3619feebdc8669e53115e5533e24707f3040926
apple-llvm-split-dir: lldb/