[testsuite] Remove testing failures vestiges.
Some tests are failing on macOS when building with the in-tree
clang, and this is because they're conditional on the version released.
Apple releases using a different versioning number, but as these are
conditional on clang < 7, they fail for clang ToT (which is 6.0).
As a general solution, we actually need either a mapping between
Apple internal release version and public ones.

That said, I discussed this with Fred , and Apple Clang 6.0 seems
to be old enough that we can remove this altogether (which means I
can delay implementing the general purpose solution for a bit).

Differential Revision:  https://reviews.llvm.org/D41101

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@320444 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 345af6861cb24f610c3afba40eb2331feca6118c)

apple-llvm-split-commit: 7fb4e100cd812a84d91a66f2a17a96c67bcd4ce5
apple-llvm-split-dir: lldb/