-gmodules: Emit debug info for implicit module imports via #include.
When a type is only used as a template parameter and that type is the
only type imported from another #include'd module, no skeleton CU for
that module is generated, so a consumer doesn't know where to find the
type definition. By emitting an import declaration, we can force a
skeleton CU to be generated for each imported module.

rdar://problem/36266156

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@321754 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 3285546c97338890987540b3f90f7e352944374e)

apple-llvm-split-commit: 028a4aaed4c40ce59946ede79208014615b3cdd1
apple-llvm-split-dir: clang/