AArch64: Cyclone: Remove SlowMisaligned128Store tuning flag
Remove FeatureSlowMisaligned128Store from cyclone flags.
This flag causes splitting of 16 byte wide stores into 2 stored of 8
bytes. This was useful on older apple CPUs which were slow for 16byte
stores that were not aligned on 16byte. As the compiler often cannot
predict the actual alignment, the splitting was choosen.

This has been a topic for a lot of debate as the splitting also
decreases performance for some benchmarks. Measuring the effects on
newer apple chips (rdar://35525421) shows that it harms more cases than
it helps. So it is time to retire this workaround.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@323289 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 68b7c2c055ede48d2d8e8c34f06ba436a0383ae0
apple-llvm-split-dir: llvm/