Compile the LLDB tests out-of-tree.
This patch is the result of a discussion on lldb-dev, see
http://lists.llvm.org/pipermail/lldb-dev/2018-January/013111.html for
background.

For each test (should be eventually: each test configuration) a
separate build directory is created and we execute

  make VPATH=$srcdir/path/to/test -C $builddir/path/to/test -f $srcdir/path/to/test/Makefile -I $srcdir/path/to/test

In order to make this work all LLDB tests need to be updated to find
the executable in the test build directory, since CWD still points at
the test's source directory, which is a requirement for unittest2.

Although we have done extensive testing, I'm expecting that this first
attempt will break a few bots. Please DO NOT HESITATE TO REVERT this
patch in order to get the bots green again. We will likely have to
iterate on this some more.

Differential Revision: https://reviews.llvm.org/D42281

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@323803 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 74e3902889d4d04cc0cb22ae9cb18dcd1046d1c2)

apple-llvm-split-commit: d6fa2b771b89612513257d9d9d07dff4cb9cb5cb
apple-llvm-split-dir: lldb/