llgs-test: Parse and store register info recieved from lldb-server
Summary:
Right now the test client is not parsing register values correctly,
which is manifesting itself in one test failing on 32-bit architectures
(pr36013). This parses the information from the qRegisterInfo packets
and stores it in the client, which will enable fixing the parsing in a
follow up commit.

I am also adding a new templated SendMessage overload, which enables one
to send a message get a parsed response in a single call.

Reviewers: eugene, davide

Subscribers: lldb-commits

Differential Revision: https://reviews.llvm.org/D43076

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@324722 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 10837fee9c04d2a53bd6966372de6ee4bd32944e
apple-llvm-split-dir: lldb/