[analyzer] NFC: Use CFG construction contexts instead of homemade lookahead.
The analyzer was relying on peeking the next CFG element during analysis
whenever it was trying to figure out what object is being constructed
by a given constructor. This information is now available in the current CFG
element in all cases that were previously supported by the analyzer,
so no complicated lookahead is necessary anymore.

No functional change intended - the context in the CFG should for now be
available if and only if it was previously discoverable via CFG lookahead.

Differential Revision: https://reviews.llvm.org/D42721


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@324800 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: b418f5b32b4dde493e31c9a9ff2e045cbbe8530e
apple-llvm-split-dir: clang/