[analyzer] Allow inlining constructors into return values.
This only affects the cfg-temporary-dtors mode - in this mode we begin inlining
constructors that are constructing function return values. These constructors
have a correct construction context since r324952.

Because temporary destructors are not only never inlined, but also don't have
the correct target region yet, this change is not entirely safe. But this
will be fixed in the subsequent commits, while this stays off behind the
cfg-temporary-dtors flag.

Lifetime extension for return values is still not modeled correctly.

Differential Revision: https://reviews.llvm.org/D42875


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@325202 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: e3aa60bd6b4bbd5538213cabd5af4436efd114ae
apple-llvm-split-dir: clang/