[analyzer] Implement path notes for temporary destructors.
Temporary destructors fire at the end of the full-expression. It is reasonable
to attach the path note for entering/leaving the temporary destructor to its
CXXBindTemporaryExpr. This would not affect lifetime-extended temporaries with
their automatic destructors which aren't temporary destructors.

The path note may be confusing in the case of destructors after elidable copy
constructors.

Differential Revision: https://reviews.llvm.org/D43144


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@325284 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: bdda180a1556a41d8e4fe32d6b38561d587bb200
apple-llvm-split-dir: clang/