[ObjC] Allow declaring __strong pointer fields in structs in Objective-C ARC mode.
Declaring __strong pointer fields in structs was not allowed in
Objective-C ARC until now because that would make the struct non-trivial
to default-initialize, copy/move, and destroy, which is not something C
was designed to do. This patch lifts that restriction.

Special functions for non-trivial C structs are synthesized that are
needed to default-initialize, copy/move, and destroy the structs and
manage the ownership of the objects the __strong pointer fields point
to. Non-trivial structs passed to functions are destructed in the callee
function.

rdar://problem/33599681

Differential Revision: https://reviews.llvm.org/D41228

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@326307 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit f4ab7b42ebfeff74222e4fe541a878f0f6a83d48)

apple-llvm-split-commit: f3dd1e6dd4e5024d1bd3af6a89feb48ae15adcb8
apple-llvm-split-dir: clang/