[test] Add dotest wrapper
This adds a wrapper around dotest, similar to llvm-lit in llvm. The
wrapper is created in the binary directory, next to LLDB and allows you
to invoke dotest without having to pass any of the configuration
arguments yourself. I think this could also be useful for re-running a
particular test case when it fails, as an alternative to "Command
Invoked".

The motivation for this is that I'd like to replace the driver part of
dotest with lit. As a first step, I'd like to have lit invoke dotest,
which would just run the complete test suite, completely identical to
what the CMake target does today. Once this is in place, we can have lit
run dotest for the different test directories, and ultimately once per
python file. Along the way we can strip out driver functionality from
dotest where appropriate.

https://reviews.llvm.org/D44002

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@326687 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 0c4d175bb29b35ba7048fbe5546dd0bf0873a80a)

apple-llvm-split-commit: d92db176161bc3a23bd4851aec6d4502d2a1db0d
apple-llvm-split-dir: lldb/