Fix standalone lldb cmake build.
The code here works in the in-tree version of the lldb cmake build. But when we
do out of tree using llvm-config, we do not have access to the clang target. If
we included llvm's targets directly via the exported llvm cmake configuration,
this would work, but for whatever reason lldb isn't doing that (I thought it
was... but maybe my memory is wrong).

apple-llvm-split-commit: 531e054d6a5ad99653ce342f72d281699d791cb7
apple-llvm-split-dir: lldb/