[analyzer] Destroy and lifetime-extend inlined function return values properly.
This patch uses the newly added CFGCXXRecordTypedCall element at the call site
of the caller to construct the return value within the callee directly into the
caller's stack frame. This way it is also capable of populating the temporary
destructor and lifetime extension maps for the temporary, which allows
temporary destructors and lifetime extension to work correctly.

This patch does not affect temporaries that were returned from conservatively
evaluated functions.

Differential Revision: https://reviews.llvm.org/D44124


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@327345 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 3af135cf37c5c166ff9814322c53bdac12292f9e
apple-llvm-split-dir: clang/