[ExpressionParser] Re-implement r327356 in a less disruptive way.
Instead of applying the sledgehammer of refusing to insert any
C++ symbol in the ASTContext, try to validate the decl if what
we have is an operator. There was other code in lldb which was
responsible for this, just not really exposed (or used) in this
codepath. Also, add a better/more comprehensive test.

<rdar://problem/35645893>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@328025 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 8f1d2e359256b6c6eb62974989fa1338241fa1f9
apple-llvm-split-dir: lldb/