[MachineOutliner] AArch64: Don't outline ADRPs with un-outlinable operands
If an ADRP appears with, say, a CPI operand, we shouldn't outline it.

This moves the check for unsafe operands so that it occurs before the special-case
for ADRPs. Also add a test for outlining ADRPs.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@328674 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 49dbb6edd824838dae26ce69e6284c1b9df6ea42)

apple-llvm-split-commit: dab7e2a71e4e4c40dbcfb9ff134f703ad87a018a
apple-llvm-split-dir: llvm/