[AArch64] Use FP to access the emergency spill slot
In the presence of variable-sized stack objects, we always picked the
base pointer when resolving frame indices if it was available.

This makes us hit an assert where we can't reach the emergency spill
slot if it's too far away from the base pointer. Since on AArch64 we
decide to place the emergency spill slot at the top of the frame, it
makes more sense to use FP to access it.

The changes here don't affect only emergency spill slots but all the
frame indices. The goal here is to try to choose between FP, BP and SP
so that we minimize the offset and avoid scavenging, or worse, asserting
when trying to access a slot allocated by the scavenger.

Previously discussed here: https://reviews.llvm.org/D40876.

Differential Revision: https://reviews.llvm.org/D45358

rdar://problem/38894780

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@329691 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: fe928c264271b8f43ab3b5e1009f911e82367098
apple-llvm-split-dir: llvm/